Split functions in main.cc to multiple headers
